// src/app/models/country-model.ts
export class CountryModel {
  'id': number;
  'name': string;
  'alpha1': string;
  'alpha2': string;
  'calling_code': number;
  'top_level_domain': string;
  'capital': string;
  'region': string;
  'currency_name': string;
  'currency_symbol': string;
  'currency_code': string
}
// export class CountryModel {
//   'name': string;
//   'alpha2Code': string;
//   'alpha3Code': string;
//   'callingCodes': number;
//   'topLevelDomain': [];
//   'capital': string;
//   'region': string;
//   'currencies': [];
// }